backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
boto3
python-dotenv
pydantic

